+++ homeassistant/components/history_stats/manifest.json
@@ -2,7 +2,7 @@
   "domain": "history_stats",
   "name": "History Stats",
   "documentation": "https://www.home-assistant.io/integrations/history_stats",
-  "dependencies": ["history"],
+  "dependencies": ["recorder"],
   "codeowners": [],
   "quality_scale": "internal",
   "iot_class": "local_polling"
