+++ tests/components/filter/test_sensor.py
@@ -81,7 +81,6 @@ async def test_chain(hass, values):
 async def test_chain_history(hass, values, missing=False):
     """Test if filter chaining works."""
     config = {
-        "history": {},
         "sensor": {
             "platform": "filter",
             "name": "test",
@@ -94,7 +93,6 @@ async def test_chain_history(hass, values, missing=False):
         },
     }
     await async_init_recorder_component(hass)
-    assert_setup_component(1, "history")
 
     t_0 = dt_util.utcnow() - timedelta(minutes=1)
     t_1 = dt_util.utcnow() - timedelta(minutes=2)
@@ -114,10 +112,10 @@ async def test_chain_history(hass, values, missing=False):
         }
 
     with patch(
-        "homeassistant.components.history.state_changes_during_period",
+        "homeassistant.components.recorder.history.state_changes_during_period",
         return_value=fake_states,
     ), patch(
-        "homeassistant.components.history.get_last_state_changes",
+        "homeassistant.components.recorder.history.get_last_state_changes",
         return_value=fake_states,
     ):
         with assert_setup_component(1, "sensor"):
@@ -208,7 +206,6 @@ async def test_chain_history_missing(hass, values):
 async def test_history_time(hass):
     """Test loading from history based on a time window."""
     config = {
-        "history": {},
         "sensor": {
             "platform": "filter",
             "name": "test",
@@ -217,7 +214,6 @@ async def test_history_time(hass):
         },
     }
     await async_init_recorder_component(hass)
-    assert_setup_component(1, "history")
 
     t_0 = dt_util.utcnow() - timedelta(minutes=1)
     t_1 = dt_util.utcnow() - timedelta(minutes=2)
@@ -231,10 +227,10 @@ async def test_history_time(hass):
         ]
     }
     with patch(
-        "homeassistant.components.history.state_changes_during_period",
+        "homeassistant.components.recorder.history.state_changes_during_period",
         return_value=fake_states,
     ), patch(
-        "homeassistant.components.history.get_last_state_changes",
+        "homeassistant.components.recorder.history.get_last_state_changes",
         return_value=fake_states,
     ):
         with assert_setup_component(1, "sensor"):
