+++ tests/components/history_stats/test_sensor.py
@@ -35,7 +35,6 @@ class TestHistoryStatsSensor(unittest.TestCase):
         """Test the history statistics sensor setup."""
         self.init_recorder()
         config = {
-            "history": {},
             "sensor": {
                 "platform": "history_stats",
                 "entity_id": "binary_sensor.test_id",
@@ -57,7 +56,6 @@ class TestHistoryStatsSensor(unittest.TestCase):
         """Test the history statistics sensor setup for multiple states."""
         self.init_recorder()
         config = {
-            "history": {},
             "sensor": {
                 "platform": "history_stats",
                 "entity_id": "binary_sensor.test_id",
@@ -146,7 +144,6 @@ class TestHistoryStatsSensor(unittest.TestCase):
         """Test when duration value is not a timedelta."""
         self.init_recorder()
         config = {
-            "history": {},
             "sensor": {
                 "platform": "history_stats",
                 "entity_id": "binary_sensor.test_id",
@@ -187,7 +184,6 @@ class TestHistoryStatsSensor(unittest.TestCase):
         """Test config when not enough arguments provided."""
         self.init_recorder()
         config = {
-            "history": {},
             "sensor": {
                 "platform": "history_stats",
                 "entity_id": "binary_sensor.test_id",
@@ -206,7 +202,6 @@ class TestHistoryStatsSensor(unittest.TestCase):
         """Test config when too many arguments provided."""
         self.init_recorder()
         config = {
-            "history": {},
             "sensor": {
                 "platform": "history_stats",
                 "entity_id": "binary_sensor.test_id",
@@ -344,9 +339,9 @@ async def test_measure_multiple(hass):
     )
 
     with patch(
-        "homeassistant.components.history.state_changes_during_period",
+        "homeassistant.components.recorder.history.state_changes_during_period",
         return_value=fake_states,
-    ), patch("homeassistant.components.history.get_state", return_value=None):
+    ), patch("homeassistant.components.recorder.history.get_state", return_value=None):
         for i in range(1, 5):
             await hass.helpers.entity_component.async_update_entity(f"sensor.sensor{i}")
         await hass.async_block_till_done()
@@ -421,9 +416,9 @@ async def async_test_measure(hass):
     )
 
     with patch(
-        "homeassistant.components.history.state_changes_during_period",
+        "homeassistant.components.recorder.history.state_changes_during_period",
         return_value=fake_states,
-    ), patch("homeassistant.components.history.get_state", return_value=None):
+    ), patch("homeassistant.components.recorder.history.get_state", return_value=None):
         for i in range(1, 5):
             await hass.helpers.entity_component.async_update_entity(f"sensor.sensor{i}")
         await hass.async_block_till_done()
